Update mundis-validator.service
It needs to run after "local-fs.target" to have all disks mounted in case of having the external disk. and also, regarding the network, it needs the online network, not only the network.

It would be better to have more delay on restart. the os will stop service after many restarts